Refactored tests to use test harness
